chore: Moved all tests to their own folder. + Changed variable name in .env. + Added a .env.example to version control. + Moved forgot password url call to .env file. + Changed dotenv call from ts to js to simplify things.
