+++ gobarber-backend/.env.example
@@ -1,0 +1,3 @@
+APP_SECRET=
+APP_WEB_URL=http://localhost:3000
+PORT=
