--- gobarber-backend/ormconfig.json
@@ -1,31 +1,0 @@
-[
-  {
-    "name": "default",
-    "type": "postgres",
-    "host": "localhost",
-    "post": 5432,
-    "username": "postgres",
-    "password": "docker",
-    "database": "gobarber",
-    "entities": [
-      "./src/modules/**/infra/typeorm/entities/*.ts"
-    ],
-    "migrations": [
-      "./src/shared/infra/typeorm/migrations/*.ts"
-    ],
-    "cli": {
-      "migrationsDir": "./src/shared/infra/typeorm/migrations"
-    }
-  },
-  {
-    "name": "mongo",
-    "type": "mongodb",
-    "host": "localhost",
-    "port": 27017,
-    "database": "gobarber",
-    "useUnifiedTopology": true,
-    "entities": [
-      "./src/modules/**/infra/typeorm/schemas/*.ts"
-    ]
-  }
-]
