+++ gobarber-backend/src/config/auth.ts
@@ -2,7 +2,7 @@ import { Secret } from 'jsonwebtoken';
 
 export default {
   jwt: {
-    secret: process.env.JWT_SECRET as Secret,
+    secret: process.env.APP_SECRET as Secret,
     expiresIn: '1d',
   },
 };
