+++ gobarber-backend/src/modules/appointments/tests/ListProvidersService.spec.ts
@@ -1,5 +1,5 @@
 import FakeUsersRepository from '@modules/users/repositories/fakes/FakeUsersRepository';
-import ListProvidersService from './ListProvidersService';
+import ListProvidersService from '../services/ListProvidersService';
 
 let fakeUsersRepository: FakeUsersRepository;
 let listProviders: ListProvidersService;
