+++ gobarber-backend/src/modules/users/services/SendForgotPasswordEmailService.ts
@@ -42,7 +42,7 @@ class SendForgotPasswordEmailService {
         file: forgotPasswordMailTemplate,
         variables: {
           name: user.name,
-          link: `http://localhost:3000/reset_password?token=${token}`,
+          link: `${process.env.APP_WEB_URL}/reset_password?token=${token}`,
         },
       },
     });
