+++ gobarber-backend/src/modules/users/tests/AuthenticateUserService.spec.ts
