+++ gobarber-backend/src/modules/users/tests/UpdateUserAvatarService.spec.ts
