+++ gobarber-backend/src/shared/infra/http/server.ts
@@ -1,8 +1,5 @@
-import dotenv from 'dotenv';
-
-dotenv.config();
-
 import 'reflect-metadata';
+import 'dotenv/config';
 
 import express, { Request, Response, NextFunction } from 'express';
 import cors from 'cors';
